Fixed bugs in SSE Type Promotion specs.

+++ lib/sparql/spec/sse/type-promotion/type-promotion-30_spec.rb
@@ -53,7 +53,7 @@ t:dateTime1		rdf:value	"2005-01-14T12:34:56"^^xsd:dateTime .
                  (t: <http://www.w3.org/2001/sw/DataAccess/tests/data/TypePromotion/tP-0#>)
                  (rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>))
           (ask
-            (filter (= (datatype (+ ?l ?r)) xsd:float)
+            (filter (= (datatype (+ ?l ?r)) xsd:decimal)
               (bgp
                 (triple t:float1 rdf:value ?l)
                 (triple t:decimal1 rdf:value ?r)
